requirements.txt
Django (1.7)
django-mptt (0.6.1)
djangorestframework (3.0.3)
nltk (3.0.1)
pbr (0.10.7)
Pillow (2.7.0)
pip (6.0)
setuptools (8.2.1)
six (1.8.0)
stevedore (1.1.0)
virtualenv (12.0)
virtualenv-clone (0.2.5)
virtualenvwrapper (4.3.1)
